[Engine] [Animation] commit skin animation

+++ Engine/Assets/Engine/Shaders/Common/common_vertex_pixel.hlsl
@@ -41,7 +41,7 @@ struct Vertex_PosUvNorTanBone
     float2 uv : TEXCOORD0;
     float3 normal : NORMAL0;
     float3 tangent : TANGENT0;
-    float4 boneIndices : BLENDINDEX;
+    uint4 boneIndices : BLENDINDEX;
     float3 boneWeights : BLENDWEIGHT;
 };
